Phase 1: extract GitHub repo browser into features/repo-browser.js
Decoupled via dependency injection: handleRepoUrl(url, { clearError, showError,
onSelectFile }) — main.js passes its URL-error helpers and handleUrl as the
file-selection handler, so the module doesn't reach back into the entry. Tests
updated to pass the hooks.

Pure refactor. main.js ~2,430 -> ~2,324 lines. Build + lint clean; 297 tests pass.
