+++ markdown-viewer/src/main.js
@@ -34,6 +34,7 @@ import {
   toggleAnnotationMode,
   renderAnnotations,
 } from './features/annotations.js';
+import { handleRepoUrl } from './features/repo-browser.js';
 
 // ===========================
 // Constants
@@ -929,7 +930,11 @@ async function handleUrl(url) {
     // Check if this is a GitHub repo URL to show the file browser
     const isRepoBrowserUrl = /^https?:\/\/github\.com\/[^/]+\/[^/]+\/?$/.test(url);
     if (isRepoBrowserUrl) {
-        const handled = await handleRepoUrl(url);
+        const handled = await handleRepoUrl(url, {
+            clearError: clearUrlError,
+            showError: showUrlError,
+            onSelectFile: handleUrl,
+        });
         if (handled) {
             if (urlInput) urlInput.value = '';
             return;
@@ -2311,117 +2316,6 @@ function applyCustomCss(cssContent) {
         document.head.appendChild(customStyleEl);
     }
     customStyleEl.textContent = cssContent || '';
-}
-
-// ===========================
-// Feature: GitHub Repo File Browser
-// ===========================
-// Enhances the URL input to accept github.com/<owner>/<repo> URLs and
-// show a list of .md files from the repo for the user to pick one.
-
-async function fetchGitHubRepoFiles(repoUrl) {
-    // Match: https://github.com/<owner>/<repo>
-    const repoPattern = /^https?:\/\/github\.com\/([^/]+)\/([^/]+)\/?$/;
-    const match = repoUrl.match(repoPattern);
-    if (!match) return null;
-
-    const owner = match[1];
-    const repo = match[2].replace(/\.git$/, '');
-
-    // Use GitHub Search API to find .md files (avoids full tree traversal)
-    const apiUrl = `https://api.github.com/search/code?q=extension:md+repo:${owner}/${repo}&per_page=100`;
-
-    try {
-        const resp = await fetch(apiUrl, {
-            headers: { Accept: 'application/vnd.github+json' }
-        });
-        if (!resp.ok) return null;
-        const data = await resp.json();
-        if (!data.items || data.items.length === 0) return [];
-
-        return data.items.map((item) => ({
-            path: item.path,
-            url: item.html_url,
-            rawUrl: `https://raw.githubusercontent.com/${owner}/${repo}/HEAD/${item.path}`
-        }));
-    } catch (e) {
-        return null;
-    }
-}
-
-async function handleRepoUrl(url) {
-    clearUrlError();
-    const files = await fetchGitHubRepoFiles(url);
-    if (files === null) {
-        // Not a repo URL or fetch failed — fall through to normal URL handling
-        return false;
-    }
-    if (files.length === 0) {
-        showUrlError('No markdown files found in this repository.');
-        return true;
-    }
-
-    showRepoBrowser(files, url);
-    return true;
-}
-
-function showRepoBrowser(files, repoUrl) {
-    let modal = document.getElementById('repo-browser-modal');
-    if (!modal) {
-        modal = document.createElement('div');
-        modal.id = 'repo-browser-modal';
-        modal.className = 'repo-browser-modal';
-        document.body.appendChild(modal);
-    }
-
-    const repoName = repoUrl.replace(/^https?:\/\/github\.com\//, '').replace(/\/$/, '');
-
-    modal.innerHTML = `
-        <div class="repo-browser-content">
-            <div class="repo-browser-header">
-                <span class="repo-browser-title">${escapeHtml(repoName)}</span>
-                <button class="repo-browser-close" title="Close">&#10005;</button>
-            </div>
-            <div class="repo-browser-search">
-                <input type="text" class="repo-browser-filter" placeholder="Filter files..." autocomplete="off">
-            </div>
-            <ul class="repo-browser-list">
-                ${files.map((f) => `
-                    <li class="repo-browser-item" data-raw-url="${escapeHtml(f.rawUrl)}">
-                        <span class="repo-file-icon">📄</span>
-                        <span class="repo-file-path">${escapeHtml(f.path)}</span>
-                    </li>
-                `).join('')}
-            </ul>
-        </div>
-    `;
-    modal.style.display = 'flex';
-
-    modal.querySelector('.repo-browser-close').addEventListener('click', () => {
-        modal.style.display = 'none';
-    });
-
-    modal.addEventListener('click', (e) => {
-        if (e.target === modal) modal.style.display = 'none';
-    });
-
-    const filterInput = modal.querySelector('.repo-browser-filter');
-    filterInput.addEventListener('input', () => {
-        const q = filterInput.value.toLowerCase();
-        modal.querySelectorAll('.repo-browser-item').forEach((item) => {
-            const path = item.querySelector('.repo-file-path').textContent.toLowerCase();
-            item.style.display = path.includes(q) ? '' : 'none';
-        });
-    });
-    filterInput.focus();
-
-    modal.querySelectorAll('.repo-browser-item').forEach((item) => {
-        item.addEventListener('click', () => {
-            const rawUrl = item.getAttribute('data-raw-url');
-            modal.style.display = 'none';
-            handleUrl(rawUrl);
-        });
-    });
 }
 
 // ===========================
